Fix services

+++ c3-checkin/src/app/services/services.component.ts
@@ -1,4 +1,4 @@
-import {Component, OnInit} from '@angular/core';
+import {AfterViewInit, Component, OnInit} from '@angular/core';
 import {Service} from '../../models/service';
 import {Member} from '../../models/member';
 import {ServicesByMember} from '../../models/servicesByMember';
@@ -15,10 +15,10 @@ export class ServicesComponent implements OnInit {
   whoNeeds = [];
   selected = [];
   services = [];
+  alreadyLoaded: any;
 
   constructor(private apiService: ApiService, private router: Router) {
     this.services = [];
-
     this.apiService.getServices()
       .subscribe((data) => {
         let i;
@@ -27,13 +27,18 @@ export class ServicesComponent implements OnInit {
           this.services.push(new Service(data[i].service_id, data[i].name));
         }
       });
-
-    this.whoNeeds = JSON.parse(sessionStorage.getItem('whoNeeds'));
-    this.selectedMember = this.whoNeeds.pop();
-    sessionStorage.setItem('whoNeeds', JSON.stringify(this.whoNeeds, null, 4));
   }
 
   ngOnInit() {
+    if (!this.alreadyLoaded) {
+      this.whoNeeds = JSON.parse(sessionStorage.getItem('whoNeeds'));
+      this.selectedMember = this.whoNeeds.pop();
+      sessionStorage.setItem('whoNeeds', JSON.stringify(this.whoNeeds, null, 4));
+
+      this.selected = [];
+      this.alreadyLoaded = true;
+    }
+
     if (this.selectedMember == null) {
       this.router.navigate(['login']);
     }
@@ -94,7 +99,8 @@ export class ServicesComponent implements OnInit {
       sessionStorage.setItem('servicesByMember', JSON.stringify(servicesByMember, null, 4));
 
       if (this.whoNeeds.length > 0) {
-        location.reload();
+        this.alreadyLoaded = false;
+        this.ngOnInit();
       } else {
         this.router.navigate(['thanks']);
       }
